Fix vscode go fmt warning

+++ common/encoding.go
@@ -11,7 +11,8 @@ import (
 type IEncoding interface {
 	// The number that uniquely identifies this encoding type.
 	Type() int32
-	WriteTo(w io.Writer) (n int64, err error)
+	//WriteTo ...
+	WriteTo(w io.Writer) (n int, err error)
 	// Read reads the contents of the encoded pixel data from the reader.
 	// This should return a new IEncoding implementation that contains
 	// the proper data.
